requirements.txt
FastAPI==0.54.1
boto3==1.13.11
python-dotenv==0.13.0
pydantic==1.5.1